feat: add DeletableChip component to improve accessibility and update Chip stories
