+++ src/components/Chip/index.ts
@@ -1,1 +1,2 @@
 export { default as Chip } from './Chip';
+export { default as DeletableChip } from './DeletableChip';
